# src/main/java/br/com/sprint4/controller/PartidoController.java
package br.com.sprint4.controller;

import br.com.sprint4.enums.Ideologia;
import br.com.sprint4.repositories.AssociadoRepository;
import br.com.sprint4.services.PartidoService;
import br.com.sprint4.services.assembler.AssociadoDTOAssembler;
import br.com.sprint4.services.assembler.PartidoDTOAssembler;
import br.com.sprint4.services.assembler.PartidoInputDisassembler;
import br.com.sprint4.services.dto.request.PartidoInputDTO;
import br.com.sprint4.services.dto.responses.AssociadoResumoResponseDTO;
import br.com.sprint4.services.dto.responses.PartidoResponseDTO;
import lombok.RequiredArgsConstructor;
import org.springframework.data.domain.Pageable;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;

import javax.validation.Valid;
import java.util.List;

@RestController
@RequestMapping
@RequiredArgsConstructor
public class PartidoController {

    private final PartidoService service;

    private final PartidoDTOAssembler assembler;

    private final AssociadoDTOAssembler associadoAssembler;

    private final PartidoInputDisassembler disassembler;

    private final AssociadoRepository associadoRepository;

    @GetMapping(value = "/partidos")
    public List<PartidoResponseDTO> listar(@RequestParam(required = false, name = "Ideologia") Ideologia ideologia, Pageable pageable){
        return service.verificacaoPartidoresponseDTO(ideologia, pageable);
    }

    @GetMapping(value = "/partidos/{partidoId}")
    public PartidoResponseDTO buscar(@PathVariable Long partidoId){
        var partido = service.buscaOuFalha(partidoId);
        return assembler.toModel(partido);
    }

    //(Listar todos os associados daquele partido)
    @GetMapping(value = "/partidos/{partidoId}/associados")
    public List<AssociadoResumoResponseDTO> listar(@PathVariable Long partidoId){
        var associados = service.listarAssociadosDeUmPartido(partidoId);
        return associadoAssembler.toCollectionModelResponse(associados);
    }

    @PostMapping(value = "/partidos")
    @ResponseStatus(HttpStatus.CREATED)
    public PartidoResponseDTO adicionar(@RequestBody @Valid PartidoInputDTO partidoInputDTO) {
        var partido = disassembler.toDomainObject(partidoInputDTO);
        partido = service.adicionar(partido);
        return assembler.toModel(partido);
    }

    @PutMapping(value = "/partidos/{partidoId}")
    public PartidoResponseDTO atualizar(@PathVariable Long partidoId, @RequestBody @Valid PartidoInputDTO partidoInputDTO){
        var partidoAtual = service.buscaOuFalha(partidoId);
        disassembler.copyToDomainObject(partidoInputDTO, partidoAtual);
        partidoAtual = service.adicionar(partidoAtual);
        return assembler.toModel(partidoAtual);
    }

    @DeleteMapping(value = "/partidos/{partidoId}")
    @ResponseStatus(HttpStatus.NO_CONTENT)
    public void remover(@PathVariable Long partidoId){
        service.excluir(partidoId);
    }

}